+++ src/arduino/SensorParty/Output.txt
@@ -1,0 +1,21 @@
+Example output:
+
+D:671,191,84,22,
+C:22.00,23.25,26.50,29.00,27.75,24.50,24.25,25.25,21.75,23.00,24.25,25.75,22.50,22.00,23.00,23.75,21.75,22.00,22.50,22.25,22.25,22.75,21.75,22.25,24.00,22.00,22.00,23.00,22.00,21.75,21.75,21.75,21.75,20.50,21.50,22.50,22.75,21.75,22.50,22.75,21.00,20.50,21.00,22.00,22.75,21.75,21.75,22.50,20.50,20.25,21.00,21.75,21.75,22.50,22.25,22.25,20.50,20.00,20.50,20.75,22.25,22.00,21.75,22.50,
+T:18.37,20.65,16.59,
+D:672,194,83,22,
+D:656,200,84,20,
+T:18.43,20.67,16.57,
+
+
+
+AMG8833 output:
+[17.50, 18.50, 21.50, 18.75, 20.75, 21.50, 20.25, 22.00, 
+17.50, 18.25, 18.50, 19.00, 20.00, 21.50, 21.00, 21.75, 
+18.25, 18.00, 17.75, 18.75, 19.50, 21.00, 20.75, 26.50, 
+17.75, 17.75, 18.50, 19.00, 19.50, 21.00, 26.50, 29.50, 
+17.25, 17.00, 17.75, 18.50, 20.00, 24.00, 29.50, 29.50, 
+18.25, 17.25, 17.50, 18.25, 21.25, 27.75, 29.75, 30.25, 
+17.50, 17.00, 17.00, 19.50, 26.25, 29.00, 29.75, 30.00, 
+17.25, 16.75, 18.00, 24.50, 28.00, 29.00, 29.50, 30.25, 
+]
