+++ src/configs/motion/cameras/video0.conf
