+++ src/configs/motion/cameras/video1.conf
