+++ src/configs/motion/cameras/video2.conf
