+++ src/configs/motion/cameras/video3.conf
