+++ src/configs/sights/minimal.json
